@@ -2,119 +2,74 @@ package chanchain
 
 import (
 	"context"
-	"reflect"
+	"sync"
 	"sync/atomic"
 	"time"
 )
 
-type Source <-chan interface{}
+type Source[T any] <-chan T
 
-func NewSource(ch interface{}) Source {
-	return Source(wrapChannel(ch))
+func NewSource[T any](ch <-chan T) Source[T] {
+	return Source[T](ch)
 }
 
-func NewTickSource(interval time.Duration) Source {
+func NewTickSource(interval time.Duration) Source[time.Time] {
 	t := time.NewTicker(interval)
 	return NewSource(t.C)
 }
 
-type Transform func(interface{}) interface{}
-
-type Chain struct {
-	v atomic.Value
-}
-
-func NewChain(ts ...Transform) *Chain {
-	var c Chain
-	c.Append(ts...)
-	return &c
-}
-
-func (c *Chain) Start(ctx context.Context, s Source) {
+// Listen receives values from Source and store it into Value
+func (s Source[T]) Listen(ctx context.Context) *Value[T] {
+	var v Value[T]
 	go func() {
 		for {
 			select {
-			case v, ok := <-s:
+			case vv, ok := <-s:
 				if !ok {
 					return
 				}
-				chain := c.v.Load().(*chain)
-				_ = (*chain)(func(v interface{}) interface{} { return v })(v)
+				v.change(vv)
 			case <-ctx.Done():
 				return
 			}
 		}
 	}()
+	return &v
 }
 
-func (c *Chain) Append(ts ...Transform) {
-	if len(ts) == 0 {
-		return
-	}
-	var t Transform
-	if len(ts) == 1 {
-		t = ts[0]
-	} else {
-		t = func(v interface{}) interface{} {
-			for _, t := range ts {
-				v = t(v)
-			}
-			return v
-		}
-	}
-	var newc chain
-	if prev, _ := c.v.Load().(*chain); prev == nil {
-		newc = func(next Transform) Transform {
-			return func(v interface{}) interface{} {
-				return next(t(v))
-			}
-		}
-	} else {
-		newc = func(next Transform) Transform {
-			return func(v interface{}) interface{} {
-				return next((*prev)(t)(v))
-			}
-		}
-	}
-	c.v.Store(&newc)
+type Value[T any] struct {
+	v  atomic.Value
+	fs []func(T)
+	mu sync.Mutex
 }
 
-type Value struct {
-	v atomic.Value
+func (v *Value[T]) Load() T {
+	r, _ := v.v.Load().(T)
+	return r
 }
 
-func NewValue(c *Chain) *Value {
-	var value Value
-	c.Append(func(v interface{}) interface{} {
-		value.v.Store(v)
-		return v
+func (v *Value[T]) change(vv T) {
+	v.v.Store(vv)
+
+	v.mu.Lock()
+	fs := v.fs
+	v.mu.Unlock()
+	if len(fs) > 0 {
+		for _, f := range fs {
+			f := f
+			go f(vv)
+		}
+	}
+}
+
+type Transform[T, U any] func(T) U
+
+func Convert[T, U any](v *Value[U], t Transform[U, T]) *Value[T] {
+	var newv Value[T]
+	v.mu.Lock()
+	v.fs = append(v.fs, func(u U) {
+		newv.change(t(u))
 	})
-	return &value
-}
-
-func (v *Value) Load() interface{} {
-	return v.v.Load()
-}
-
-type chain func(Transform) Transform
-
-func wrapChannel(ch interface{}) <-chan interface{} {
-	t := reflect.TypeOf(ch)
-	if t.Kind() != reflect.Chan || t.ChanDir()&reflect.RecvDir == 0 {
-		panic("channels: input to Wrap must be readable channel")
-	}
-	realChan := make(chan interface{})
-
-	go func() {
-		v := reflect.ValueOf(ch)
-		for {
-			x, ok := v.Recv()
-			if !ok {
-				close(realChan)
-				return
-			}
-			realChan <- x.Interface()
-		}
-	}()
-	return realChan
+	v.mu.Unlock()
+	return &newv
 }
